Feat: password settings (#130)
* wip: password settings

* wip: password settings

* wip: password settings

* wip: password settings

* fix: review comments

* fix: review comments

* fix: review comments

* fix: review comments
